Optimize move generation (#9)
* Store masks, magics, shifts and attacks in a single lookup table

* Speed up king move generation by removing make/unmake move

* Refactor generateCastlingMoves

* Refactor Bitwise, PinCalculator and RayCalculator

* Use int for Move instead of short - gain from decreased memory does not outweigh cost of constant casting to int

* Run perft tests

* Add comments

+++ src/main/java/com/kelseyde/calvin/board/Move.java
@@ -7,12 +7,15 @@ import java.util.Optional;
 
 /**
  * Represents a single chess move.
- * All the information required to represent a move can be encoded in 16 bits.
+ * All the information required to represent a move is encoded in a 16-bit integer value field.
  * <p>
- * bit 0 - 5: start square (0 - 63)
- * bit 6 - 11: end square (0 - 63)
- * bit 12 - 15: special moves (promotion, castling, pawn double moves, en passant)
- * @see <a href="https://www.chessprogramming.org/Encoding_Moves">Chess Programming Wiki</a>
+ * The encoding of the move information in the value field is as follows:
+ * <ul>
+ *     <li>Bits 0 - 5: Start square (0 - 63)</li>
+ *     <li>Bits 6 - 11: End square (0 - 63)</li>
+ *     <li>Bits 12 - 15: Special move flags (promotion, castling, pawn double moves, en passant)</li>
+ * </ul>
+ * For more information on move encoding, see <a href="https://www.chessprogramming.org/Encoding_Moves">Chess Programming Wiki</a>.
  * <p>
  * Largely inspired by Sebastian Lague's Chess Coding Adventure:
  * @see <a href="https://github.com/SebLague/Chess-Coding-Adventure">Chess Coding Adventure</a>
@@ -21,6 +24,7 @@ import java.util.Optional;
 @EqualsAndHashCode
 public class Move {
 
+    // Special move flags
     public static final short NO_FLAG = 0b0000;
     public static final short EN_PASSANT_FLAG = 0b0001;
     public static final short CASTLE_FLAG = 0b0010;
@@ -30,31 +34,65 @@ public class Move {
     public static final short PROMOTE_TO_ROOK_FLAG = 0b0110;
     public static final short PROMOTE_TO_BISHOP_FLAG = 0b0111;
 
+    // Masks for extracting start and end squares from the value field
     public static final int START_SQUARE_MASK = 0b0000000000111111;
     public static final int END_SQUARE_MASK = 0b0000111111000000;
 
-    private final short value;
+    private final int value; // 16-bit integer value representing the move
 
-    public Move(short value) {
+    /**
+     * Constructs a Move instance with the given encoded value.
+     *
+     * @param value The encoded value representing the move.
+     */
+    public Move(int value) {
         this.value = value;
     }
 
+    /**
+     * Constructs a Move instance from the given start square and end square.
+     *
+     * @param startSquare The starting square of the move (0 - 63).
+     * @param endSquare   The ending square of the move (0 - 63).
+     */
     public Move(int startSquare, int endSquare) {
-        this.value = (short) (startSquare | endSquare << 6);
+        this.value = (startSquare | endSquare << 6);
     }
 
-    public Move(int startSquare, int endSquare, short flag) {
-        this.value = (short) (startSquare | endSquare << 6 | flag << 12);
+    /**
+     * Constructs a Move instance from the given start square, end square, and move flag.
+     *
+     * @param startSquare The starting square of the move (0 - 63).
+     * @param endSquare   The ending square of the move (0 - 63).
+     * @param flag        The special move flag representing additional move information.
+     */
+    public Move(int startSquare, int endSquare, int flag) {
+        this.value = startSquare | (endSquare << 6) | (flag << 12);
     }
 
+    /**
+     * Retrieves the starting square of the move.
+     *
+     * @return The starting square of the move (0 - 63).
+     */
     public int getStartSquare() {
         return value & START_SQUARE_MASK;
     }
 
+    /**
+     * Retrieves the ending square of the move.
+     *
+     * @return The ending square of the move (0 - 63).
+     */
     public int getEndSquare() {
         return (value & END_SQUARE_MASK) >>> 6;
     }
 
+    /**
+     * Retrieves the piece type to which a pawn is promoted in this move, if applicable.
+     *
+     * @return The piece type to which a pawn is promoted, or null if no promotion occurs.
+     */
     public Piece getPromotionPieceType() {
         return switch (value >>> 12) {
             case PROMOTE_TO_QUEEN_FLAG -> Piece.QUEEN;
@@ -65,6 +103,12 @@ public class Move {
         };
     }
 
+    /**
+     * Retrieves the special move flag representing additional move information.
+     *
+     * @param pieceType The type of piece to which a pawn is promoted.
+     * @return The special move flag indicating pawn promotion.
+     */
     public static short getPromotionFlag(Piece pieceType) {
         if (pieceType == null) {
             return NO_FLAG;
@@ -78,22 +122,48 @@ public class Move {
         };
     }
 
+    /**
+     * Checks if this move represents a pawn promotion.
+     *
+     * @return True if the move is a pawn promotion, false otherwise.
+     */
     public boolean isPromotion() {
         return (value >>> 12) >= PROMOTE_TO_QUEEN_FLAG;
     }
 
+    /**
+     * Checks if this move represents an en passant capture.
+     *
+     * @return True if the move is an en passant capture, false otherwise.
+     */
     public boolean isEnPassant() {
         return (value >>> 12) == EN_PASSANT_FLAG;
     }
 
+    /**
+     * Checks if this move represents castling.
+     *
+     * @return True if the move is a castling move, false otherwise.
+     */
     public boolean isCastling() {
         return (value >>> 12) == CASTLE_FLAG;
     }
 
+    /**
+     * Checks if this move represents a pawn double move.
+     *
+     * @return True if the move is a pawn double move, false otherwise.
+     */
     public boolean isPawnDoubleMove() {
         return (value >>> 12) == PAWN_DOUBLE_MOVE_FLAG;
     }
 
+    /**
+     * Checks if this move matches another move.
+     *
+     * @param move The move to compare against.
+     * @return True if the moves match, false otherwise.
+     */
     public boolean matches(Move move) {
         if (move == null) return false;
         boolean squareMatch = getStartSquare() == move.getStartSquare() && getEndSquare() == move.getEndSquare();
@@ -102,5 +172,4 @@ public class Move {
                 .orElse(true);
         return squareMatch && promotionMatch;
     }
-
 }
